Removed presumably redundant/outdated attributes.
This is a bit out of thin air, but both attributes _short_name_ and _type_ are only listed for _Direct Connect_ and are implied in the resource name and format respectively.

While at it, I've rearranged the attribute order to match the apparently canonical one in place with all other services.

Also, I've replaced the apparently accidental tabs sprinkled in there (sorry for the resulting diff readability, but it seemed odd to split that into a separate pull request).

Finally I've adjusted the name from _AWS Direct Connect Overture Service_ to just _Direct Connect_, i.e. matching the official name without the AWS/Amazon prefix, similar to all other services.