Remove get_auth function
Endpoint was the only place calling auth.get_auth so I've moved the
functionality into the endpoint module.  Instead of doing an if/else
chain, I replaced that functionality with a dictionary of auth name
to auth class.  This will make it easier for plugins/event hooks to
override this dictionary to provide their own custom auth.