Retry ClosedPoolError from urllib3
This is a workaround for https://github.com/shazow/urllib3/pull/204
This is fixed in requests 2.0.0, but we'll have to evaluate whether
or not we want to upgrade.

In the meantime, the workaround for the above issue is to just retry
if we encouter that error.  By that time the poolmanager has the
connection pool primed, so a retry attempt will grab the connection
pool from the cache, bypassing the race condition.

This issue surfaces when trying to make multiple service calls
in parallel (for example `aws s3 cp`).