Reset the stream on retry
When we retry a request, we first need to rewind the stream
if possible.  Otherwise, we will only send a portion (or none)
of the stream which will cause the server to wait because our
content length won't match what we've sent.  Eventually the server
will just close out the connection.

Fixes aws/aws-cli#401.