Adds necessary copyright line to license.
Adds Amazon, Inc. to copyright.

Corrects statements claiming the project uses the "New" BSD license.
The text in LICENSE.txt is from the MIT, which is compatible but not
identical; see http://opensource.org/licenses/MIT and compare with
http://opensource.org/licenses/BSD-3-Clause

Corrects license directive in setup.py to explicitly state the license
instead of giving the contents of LICENSE.txt.