Allow loader *instance* to be passed in
A loader instance (instead of a loader class) should be passed
to the session.  By requiring a loader class, this ties the __init__
to the interface of the loader, which should not be the case.  This
also makes it harder to share loader instances (which is where any
kind of caching would live) across sessions.

As a result of this I updated the unit tests to remove all the
patch_session code, and use the *public* API to set the loaders.

Additionally, many of these tests were written before BaseSessionTest
was around, so in many cases, the tests could just be updated to use
BaseSessionTest and use the pre-created session instance created
in the base class's setUp.