Use get_config_variable instead of profile attr
session.get_config_variable('profile') != session.profile.

In the case of "--profile", we set the session.profile attr,
but in the case of the env var, we look it up when we create
the scoped config or when we call get_config_variable('profile').

I think this can probably be fixed such that we always set
session.profile, but this fixes the immediate issue of
aws/aws-cli#805.