Revert "Merge branch 'expect-100' into develop"
The expect-100 changes do not work in python2.6.

The internal structure of HTTPConnection is different
in python2.6 so this needs to be reworked.

I'm still fairly confident that this can be done in python2.6,
it just needs additional research and time.

This reverts commit a23770e72758674a4274d68182f6a99d83810880, reversing
changes made to 13e8eecf9a1f887ca4e55de4d3c92aedb88deec5.