Recreate failed requests to refresh timestamp
If the request timestamp is not updated before a retry, it may
eventually fail with a RequestTimeTooSkewed error. Create a new request
before retrying so that the timestamp is updated. With a new request,
resetting the stream is no longer necessary, but the existing response
needs to be closed so the server stops waiting for it to be handled.

This requires some refactoring to pass through the original request
parameters and signer to _send_request(). An existing test using
_send_request() needs to be changed to reflect the new signature.