Don't parse response body when output shape is None
Fixes #352.  To ensure no other protocols were susceptible,
I reorganized part of the response parsing unit tests such that
testing for output shapes of None are grouped in a single class.

I've also added an integration test to ensure this functionality
with an actual AWS service.  This helps ensure we don't regress
in the future.