Handle case where only type key is defined in waiters
There's an issue where the acceptor type is defined in __defaults__
but no other values for failure is defined.  This causes undefined
behavior for us and results in any waiter config that has an
acceptor_type defined in a parent config but no explicit failure_*
values throwing an exception when used.

The fix is not ignore these specific waiter configs.