Fix CopySnapshot handler
This was previously crashing because the event interface was updated
to remove the operation object being sent as a kwarg.  This commit
adds this back.

To ensure we don't regress on this again, I've added an integration
test that actually copies a snapshot (encrypted and not encrypted).
This will ensure that this functionality will not be lost despite
any changes we might make to handler interfaces.