Implement waiters v2 spec
This commit implements the new waiters v2 spec, which is a
simplified format from the previous version.  The implementation
is much easier to follow, and also does not require any model
translations, we can consume the model directly without any
changes.

The Service and Client objects have been updated to use this new
interface.

There will be a separate commit to convert the waiters to the new
v2 format.