Add back support for sigv4 overrides
This allows you to change the signature version used when
signing requests for a particular service.

This isn't strictly needed to communicate with various endpoints,
any region specific signature versions are captured in _endpoints.json.
However, with S3 we sometimes need to switch to sigv4 even if it
otherwise wouldn't be required.