Revert "Merge branch 'per-client-events' into develop"
This reverts commit 4dd86730fdad5e657ba6968f7976a79113096708, reversing
changes made to 5c636ab1064b475e7b00e63d2bd1b170bc43b89d.

This is causing test failures in the integration tests.
It appears to happen in the deepcopy code when we encounter
a partial object.

We also don't encounter a partial object unless we also use
a session to call .get_service(), which registers additional handlers.
This will require some investigation so I'm reverting in the meantime.