Fix sporadic auth test failure
This was because datetime.utcnow() is called when signing the request,
and if there was a secondary boundary between when the two requests
were signed you'll get a different signature.  The fix is to just patch
out the current time so it always returns the same time for the duration
of the tests.