Disable auth rebuilding on redirects
In newer versions of requests, the Authorization
header is removed if a redirect occurs where the new
host does not equal the original host.  In the case
of S3 redirects, we know that this is ok so we need
to add back the old behavior.  We do this by subclassing
the session and overriding its rebuild_auth() class to
be a no-op.  That is, it just keeps the auth from the
original request.