Handle urllib3 timeout errors
This is an attempt to fix an issue where a successful HTTP request is made
but while reading the raw low-level response we may get a `urllib3` error
raised instead of the corresponding `requests` error.

The error presents itself when sending/receiving large files to/from S3
with the AWS CLI when on a bad network connection.

This makes sure that our retry handler is capable of handling this case
and adds a test for it.