Remove region name throughout endpoint
There is now only one place where the region name is passed to any class
in endpoint.py and that is with EndpointCreator's create_endpoint. If you
want to create and endpoint from scratch you need the region to help resolve
the host. However, any of the region deciding logic is moved out of the
Endpoint class and into the client class. The get_endpoint and
get_endpoint_complex public methods were removed to encourage use through
the endpoint creator as the Endpoint class is considered an internal class.