Standardize on AWS_ prefix for env vars
Given that boto3 is using botocore and botocore is close to GA,
we've decided to just standardize on the default values of
AWS_ instead of BOTO_.  This gives better interop between SDKs.

If users need the old behavior they can always pass in an
env var mapping that uses the old values.