Update and simplify botocore loaders to new dir layout
This reworks and greatly simplifies the loaders used in botocore.
This is part paying off technical debt as well as standardizing
on a shared common format.

In terms of the new layout, the loaders docstring discusses how it
looks like, but we now support the directory layout used by the
other AWS SDKs.

In terms of the loaders themselves:

* Remove any fuzzy API version matching, you either can specify an
  exact API version or you can get the latest API version.
* Move up the service type names ('service-2', 'paginators-1', etc)
  out of the loader.  The session/client objects should know about
  those details, but not the loader.
* Remove all file extensions from the loader class.  This actually
  allows pluggable loader types besides JSON.  This wasn't
  actually possible previously.

As part of this change, the type_name is now required in loader method
calls.

This moves the responsibility of knowing what service
model type is being loaded up one level (either to the
session of the client).