Simplify retry configuration
This changes two things:

* Move the ``throttling_exception`` rule up to the default errors to retry
* Move the ``throttling`` rule up to the default errors to retry

This came about because the ``workspaces`` service would receive
400/ThrottlingException errors that were not being retried.

The simple fix would be to add this missing condition to a new
``workspaces`` entry.

However, the error can occur for any json or rest-json service and
we'd need to keep this up to date every time we added support for a new
json/rest-json service.  Same thing with the 400/Throttling exception.
This will happen for any query service.  Moving this up to the default
set of errors to retry means we won't have to keep updating this as new
errors are added.

It also has the benefit that the remaining per-service conditions are
just the special cases.