Fix pagination issue
The issue is we need to stop paginating if we recieve an empty string for
a next token marker. Before, if we did not have a check to see that the next
token was not the same as the previous token and throw an error, we would have
hit an infinite loop.