Add option to force s3 virtual host addressing
You can force virtual style addressing through the config file:
s3 =
    addresing_style = virtual_host

You can also specify it in the Config object passed in when a client
is instantiated. The virtual host addressing no longer defaults to the
s3 global endpoint. Instead, it will use the endpoint url determined from
the region or if the endpoint url that is specified upon instantaition of
a client.