Special case versionId query param in urlencoding
A while back there was a bug fix that did not url encode
the query string.  This was so customers could provide
a ?versionId to indicate they wanted a specific version
copied (https://github.com/boto/botocore/issues/509).

However, it's possible that a customer's key name ends
with something that looks like a query string, e.g.
"foo?a=b&c=d".

This modifies this code to only skip the uri encoding
if there's a query string with versionId.

There's still one edge case.  It's ambiguous for a key
such as 'foo?versionId=bar' whether you meant the entire
string to be the key name or whether you meant
key=foo, versionId=bar.  In this case, we assume you meant
the latter, so it's not possible to copy a key with
'?versionId=...'.

In order to fix this edge case we'd need to support some
explicit way of the customer specifying the bucket, key,
and versionId.