Refactored paginators to use a dict rather than a custom seperator.
Previously, our custom pagination token came in the form:
`serviceToken___5000` where serviceToken represented the service's
pagination token(s) and the number at the end represented where in the
array we should slice the response. The problem is that a token for
dynamodb or s3 can include a triple underscore. So instead of seeing one
token and an index, we see two tokens and an index. This causes the
paginator to fall on its face.

The solution is to use a dictionary instead, which brings us in line
with how both Ruby and PHP do pagination. So our token will now look
like `{"ServiceToken": "value", "truncate_amount": 5000}`. In addtion
to dumping it to JSON, we will also base64 encode/decode to obfuscate
how it is built, since customers are not supposed to be able to inject
their own custom token. The JSON encoding / decoding gives us the
ability to seamlessly use tokens which are not themselves strings, such
as dynamodb's `ExclusiveStartKey`.