Handle case when s3 config key is not a dict
This fixes a regression for the error case when a user
has an invalid CLI config file.  Examples include:

```
[default]
s3 = foo

[profile foo]
s3 =
signature_version = s3v4
```

Previously, we would check that the 's3' key was a dict type
before trying to access values.  This puts back the old
behavior.

Long term, we're going to need something that properly validates
the config keys and gives good error messages for these conditions.