Remove Transfer-Encoding if Content-Length is set.
The requests library will set the Transfer-Encoding header if the
following is true:
1. the data object is a stream (has __iter__() defined)
2. is not a list, tuple, dict, or a string
3. the length cannot be determined

Notably, a user-supplied Content-Length header is ignored and there
is no knob to set Transfer-Encoding.

As AWS S3 version 2 signer does not support chunked Transfer-Encoding,
boto3 needs to be careful about propagating the Transfer-Encoding
flag. This patch removes the Transfer-Encoding header if
Content-Length is set by the user (even if length of the data cannot
be determined).

Fixes #911