Remove docutil dependency from setup.py
Docutil is used only in the following places:

* botocore/docs/bcdoc/textwriter.py

And is also referenced in:
* setup.py
* setup.cfg
* requirements.txt

Since during the setup.py execution the documentation is not generated,
I think we should not require docutil to be installed. We could create a
secondary target to create the documentation, but this is a different
problem/opportunity/situation.