Switch to codecov for coverage reports
This required some changes to work around the fact
that we run our tests from the tests/ directory in order
to ensure we import from the installed .whl file.  This
means that our .coveragerc file is not being sourced.

To fix that, I copy the .coverage data file to the top level
diretory, delete the coverage.xml file (which includes vendored files
we want to ignore), and let codecov regenerate the coverage.xml file.

I had to make a small change to the .coveragerc file to ignore
the botocore vendored packages.