Replace chars we can't decode with placeholder char
It's possible that the console output from EC2 contains
non string chars we can't utf-8 decode.  When that happens
we were previously just leaving the input untouched, i.e
base64 encoded.

This made for an awkward interface for customers.  We couldn't guarantee
whether or not the content was base64 decoded because it would depend
on the actual ec2 console output content.

This change attempts to make things more consistent by
replacing any chars we can't decode with the unicode
`\ufffd` char (the unicode replace char, i.e the question mark).

The one downside with this approach is that it is possible
to sometimes not get the unaltered contents of the console
output because chars have been replaced with `\ufffd`.
We'll need to weigh whether or not we think that's worth
the benefit.  Given this content is usually text content,
this seems like a reasonable compromise.