Use weakref for client in S3 redirector
The S3RegionRedirector holds a reference to the client, which was
causing memory leaks on python 2.6. We didn't notice this previously
because our tests give us a certain tolerance and we were still
falling under that tolerance. If more events are added to the
event emitter by default, it can cause the amount of memory leaked
to go past our tolerance.

This fixes the issue by making the S3 redirector hold a weakref to
the client instead of a strong reference. We don't have to worry
about it being GC'd because customers will be holding a strong ref
to the client for us.