Extract client arg construction to separate module
This commit is a pure refactoring (no functional changes)
that extracts all the logic for creating client_args into
a separate module and class.

The motivation for this change is to separate the process
of constructing a client (the logic of building API methods
and class construction via type()) with the logic for figuring
out the args to pass to a client.  These are separable concerns
so it makes sense to keep this out of the client builder.

There's a number of improvements that can be made to this class,
but this initial refactoring was strictly about keeping the same
behavior while extracting out the code.  My plan in the
future is to make this class more functional and move all of
the side effects out to a separate class.  This will make it
much easier to unit test.

The unit tests required minimal changes (just had to change a few
import strings used for mock.patch).