Remove unused code from translate.py
This was a module that was heavily used in earlier versions
of botocore.  Most of the code is now no longer used with the exception
of retry code.

There was quite a bit of useful validation we used to have for
pagination code that I've converted to a functional test in a separate
commit.

I also double checked that nothing in botocore, aws-cli, and boto3.
is using the translate module outside of the retry config translations.