Add ability to use member name for string values in structs
While using an empty string is useful for input shapes, if we
want to generate a sample output shape, it's nice to have
something better than an empty string.  In this case, using
the struct member name is enough.

This is also helpful in validating JMESPath expressions in waiters
because we can don't have to special case empty strings being
False.