Enforce GMT == UTC
In the UK the time switches from GMT to BST and back as part of
their daylight savings time. time.tzname will therefore report
both time zones. dateutil sees that the time zone is a local time
zone and so parses it as local time, but it ends up being BST
instead of GMT. To remedy this issue we can provide a time zone
context which will enforce GMT == UTC.