Lazy load the exceptions propery on the client
Creating the exceptions property takes a significant amount
of time; slows down the client creation by three fold if created
upfront. This is because every error in every operation needs to
be traversed and a class needs to be created for each one. So
instead only fully load the exceptions property if accessed on
the client. With this change, client creation time is unaffected.

Also made optimizations to minimize the number of new exception
classes that get created when it does get loaded.