Ensure no creds are automatically picked up in functional tests
There's a failing RDS functional test.  It looks like this failed
because travis CI does not have any credentials configured (which is
expected).  However, because credentials are configured on the
dev laptops, you won't see this failure locally.

To fix this issue, this commit ensures that all the default credentials
won't be picked up.  This help brings the travis environment closer
to a dev environment.

This this commit I now see the failing RDS test locally, even if I
have ~/.aws/config or ~/.aws/credentials configured.