Reuse exception classes across clients
Allows users to catch an exception from one client with any other
client of the same service. To do this the error factory needed
to be hoisted to the session and injected into the client creator
so that all client's from the same session share the same error
factory cache.