Add keyring support to config credential provider
This will automatically draw the AWS secret key from your keyring if
available and configured to do so. If configured to do so but the
keyring is not available, an exception will be thrown.

Support for the old boto-specific config file is not being added at
this stage, due to the documentation recommending the newer AWS config
file be used instead.