Make links anonymous
Sphinx will throw warnings if you have two uri's with the same name
but different targets because it helpfully assumes your intent was
to make an embedded uri global. This add an additional underscore to
the end of those links to make Sphinx treat them as 'anonymous'.

The links do still work properly without this change, but the CLI
will spit out warnings without it, which is not a great customer
experience.

Fixes #1268
Fixes aws/awscli#2786