Disable proxy config when fetching container credentials
This ensures we do not look at proxy config even if
the HTTP(S)_PROXY env var is set.

This happens if we create the requests.Session for the user.
If a user provides a requests.Session (which isn't the default
behavior) then we do not mutate the provided session.

See: http://docs.python-requests.org/en/master/api/?highlight=trust_env#request-sessions