Fix BytesWarning in parsers.py
In Python3, When Python is started with the -b flag, botocore issues a
warning:

  BytesWarning: str() on a bytes instance

As response['body'] is a bytes object, can't coerce to str without
decoding. This warning can indicate mishandling of bytes/str boundaries.
As the value is used for logging purposes, simply log the raw response
body by using the %r placeholder.

When Python is started with the -bb flag, this warning becomes an error.
For additional details on the flag, see:

https://docs.python.org/3/using/cmdline.html#miscellaneous-options