Simplify xform_name() to not separate words by numbers
The _number_cap regex has proven to not be the "right thing"
to do in practice.  In practice we end up having to put
customizations in to fix what the _number_cap regex is doing.
The most common cases are (left being what it does, right being
what we actually want):

* `s_3 -> s3`
* `ec_2 -> ec2`
* `ipv_6 -> ipv6`

This change removes the _number_cap regex, which in turn removes
our partial renames hack and simplifies the code.

This ended up having no impact for botocore (aside from one
unit test I had to fix), I verified that all the names we
`xform_name()` on in botocore are identical between this change
and the latest develop branch.  Most of these changes affect
the parameters of operations, and in botocore you pass the
parameter exactly as its cased in the service model.

I will have to make changes in the CLI though.  Because the
CLI transforms top level parameters, I'll have to go through
and set aliases for all the parameters that are changed
because of this (about 10 parameters).