Pass headers into presigned_url
Some operations that require a body must have a signature of that body
with which to sign the url; otherwise the body may be changed after the
url has been signed. Passing headers into `generate_presigned_url`
allows the signature to be correctly generated with the header that
provides a guarantee that the body of the request has not been changed.