Refactor configuration providers
Removed the general purpose config value providers in favor of more
specific providers for each step in our configuration chain. This makes
it a bit clearer what exactly is going on at each step of the config
chain since each class is specific to the task it is performing.

The `logical_name` was removed from the signature
of the `provide` method on the provider classes as well.

In addition, a number of classes were renamed for clarity, and some misc
logic was cleaned up to be a bit cleaner as well.