Add the SESSION_VARIABLES back to the session object
The session.session_vars_map was a publicly accessible map that used to
have tuples that configured how variables were loaded, this map was
removed in favor of a new system. The new system is still in place but
the tuples have been put back so that they can be read from the session.