Add generated paginators
This adds generated paginators for a number of services. This only adds
the simple cases where the response contains only two elements: a list
which is the result key and the next token.

The script that generated these will look for the standard values
('NextToken', 'MaxResults') but also look for any values that are common
within a given service.