Don't throw error for invalid old starting token
Botocore has support for its own starting token formats, the older
of which simply splits up necessary data with triple underscores.
This format has been deprecated for a while due to various
shortcomings, but there's still input support for it for backwards
compatibility. The last element in that token format should be an
integer, which indicates how far to slice into the response. The
problem is that some services have tokens which include triple
underscores, so we think it's that format but malformed. This fixes
the problem by passing the full starting token along if parsing
fails.

Fixes boto/boto3#1984