Add AssumeRoleWithWebIdentity credential provider
The web identity provider uses the full config off of the session
instead of relying on the session's get_config. While the new config
system in session is nice it has two major probles that make it
unsuitable for this use case:

1) If a profile is statically configured we need to ignore the
environment variables for that configuration. While it is possible to
use the methods kwarg, it's extremely inefficient.
2) The larger blocker is that this provider needs to support loading
configuration from a profile that is not the currently configured
profile for a session (e.g. when used as a source_profile in an assume
role chain). This is not trivial to add to the session's get_config and
doesn't really fit into the way it is designed.

Using the full config is what the other profile based providers do,
and this makes this new provider consistent with that.

* Add retries to sts for IDPCommunicationError

* Backport the StubbedSession class from chalice

This backports the StubbedSession class from chalice and refactors some
of the assume role tests to use as it simplifies things. This does not
backport the improvements or wrapper class that chalice has for the
client level stubber.