Support process profiles in assume role
This adds support for profiles that use the credential process as source
profiles for assume role. Since profile-based providers take the profile
name as a constructor argument, partials are used to fill out everything
else. The assume role provider will then construct a new credential
resolver on the fly and use that to get the credentials from the
profile.

A better, but backwards incompatible, solution would be to:

A. Have the credential resolver implement the credential provider
interface
B. Have the profile-based providers take profile as a load argument
C. Have all of the profile-based providers in their own sub-chain
D. Give a reference to that chain to the assume role provider

This would be nice to have for a V2