Proxy explicit regions to the credential resolver
Currently the credential resolver would use the region provided
via env vars, config file, or the session. However, it is possible
that region is only provided when creating the client. For these
cases, we want to use the region provided to the client when
creating clients used to resolve credentials.