Remove vendored Requests
Documentation says that vendored versions of requests and urllib3
are no longer being used and have been replaced with
a direct dependency on upstream urllib3 and requests
is no longer a dependency of botocore.
https://botocore.amazonaws.com/v1/documentation/api/latest/index.html#upgrading-to-1-11-0

However, the "requests is no longer a dependency of botocore"
statement is not true currently, since requests are used
in botocore/exceptions.py and in tests.

I believe it's time to remove vendored requests.