Update waiter test to assert for None
Assert something is Falsy is not specific enough. If the waiter
is doing a ``length`` check, it is possible for ``False`` to be
returned back. Really what we want to be checking here is that
``None`` is not returned back as that indicates the JMESPath
expression did not find anything that matched.