Allow stderr passthrough for credential_process
Many organizations now require MFA for AWS login, or perhaps fetch
credentials from a command-line password manager. While `stdout` of
the `credential_process` is used to acquire the credential token,
`stderr` is captured for raising an exception back to the user.

In order enable user-interaction with the credential provider process,
it must have access to some stream the user can see: in this case,
`stderr`. This commit provides that change.

In order to maintain a relatively high level of backwards
compatibility, we will continue to capture `stderr` for raising
exceptions _unless_ `stderr` is reporting as being a TTY device, in
which case we assume that user-interaction is more favorable than
detailed exceptions.