Use explicit aws-global region for dualstack
This ensures that we follow the same behavior for any
psuedo-regions when formatting a dualstack endpoint where the
actual region name is used in formatting the dualstack endpoint.
If this is not done and the user explicitly sets aws-global,
the request would be sent to us-east-1 regional endpoint,
which is not necessarily the same.