Handle top level Error key for non-error responses
The `standard` retry mode will attempt to introspect any parsed response
for a top level `"Error"` key. If present the standard retry strategy
assumes that the response is an error response and that there will be a
nested dictionary under the `"Code"` key. This assumption is not
generally safe as standard responses can also contain a top level
`"Error"` key.