Add CRT auth package
This commit pulls in CRT auth abstractions from the develop branch in order to
have a starting to point to pull in future updates related to CRT
auth. Note the code does not actually use any of the CRT
auth abstractions when making API requests. The main difference
in porting the package is that because CRT is a required dependency
there is no branching logic for conditionally using CRT auth. Eventually, we
should remove the pure-Python sigv4 auth abstractions and completely
rely on the CRT auth abstractions for sigv4 signing.