fix-credential-process-without-stdin
in the ProcessProvider class, the _retrieve_credentials_using method uses the self._popen to open a subprocess. If there is no stdin attached to currently running program, then this call can result in an "handle is invalid" error on Windows. Adding the parameter stdin=subprocess.PIPE fixes the issue for this case.