Remove repo root from path when running tests
Both nose and py.test will add the first parent directory it
encounters that does not have a __init__.py to the sys.path. In
our case, this is the root of the repository. This means that Python
will import the botocore package from source instead of the installed
wheel distribution. So the new environment variable provides
the option to remove the repository root from sys.path to be able to
rely on the installed distribution when running the tests.

It is also worth noting that nose has --no-path-adjustment to fix
this, but py.test does not have an option available without large
refactorings of the tests.