Copy env vars when injecting fake creds for test
Without providing a copy of the current environment variables,
the cmd-runner may not work as we are not proxying over environment
variables that are needed to function correctly. For example, the
environment can be setting the PYTHONPATH in order to find
additional Python packages that are not installed in the default
site packages.

The code is copied from https://github.com/aws/aws-cli/pull/6725

Fixes https://github.com/boto/botocore/issues/2676