Add Type field for query errors from JSON protocol
By default, the query parser parses all fields from the error
response and includes them in the parsed error dictionary. This
included a `Type` key. However, JSON protocol based services never
return a `Type` as part of the parsed error response. So for
backwards compatibility purposes, the SDK needs to parse the
`Type` key from the x-amz-query-error header and include it as
part of the parsed error response when the header is found for
JSON protocol based services.