Shallow copy config value store and providers
Instead of deep copying, use a shallow copy of the config value store
when using it for creating a new client from a session. Only use a deep
copy for updating the section provider.

Update behavior of the smart defaults update functions to use a copy
of the existing config providers in all cases. Update the logic for
determining which new provider to create.

Move unit tests for config provider smart defaults to functional tests
and rework them to better test behavior. A mocked defaults data file is
used to confirm that the correct values are loaded to protect against
changes to the live data file.