Implement configured endpoint URL resolution
Adds a config provider to resolve the endpoint URL provided in an
environment variable or shared configuration file. It resolves the
endpoint in the following manner:

1. The value provided through the `endpoint_url` parameter provided to
   the client constructor.
2. The value provided by a service-specific environment variable.
3. The value provided by the global endpoint environment variable
   (`AWS_ENDPOINT_URL`).
4. The value provided by a service-specific parameter from a services
   definition section in the shared configuration file.
5. The value provided by the global parameter from a services definition
   section in the shared configuration file.
6. The value resolved when no configured endpoint URL is provided.

The endpoint config provider uses the client name (name used to
instantiate a client object) for construction and add to the config
value store. This uses multiple lookups to handle service name changes
for backwards compatibility.