Add functional tests for configured endpoint url resolution
The tests use a data file to load tests that enumerate the ways that the
endpoint URL can be defined and asserts that the correct endpoint is
used in a request and that it is ignored correctly if the appropriate
shared config file property or environment variable is supplied.

They also assert that the correct config section name and environment
variable name are used to configure and override the endpoint URL for
every AWS service.