Fix calculation of truncated binary exponential backoff
The intent of the jitter in the exponential backoff calculation is to
create a uniform random distribution.  However, the current calculation
applies the truncation after the randomization.  This results in
clamped/truncated distribution which is not what we want.

This only affects cases where the calculated backoff exceeds the
`MAX_BACKOFF`, and is more pronounced the more you exceed `MAX_BACKOFF`.
The current default of 3 max attempts is well below this value so this
change has no effect on the default behavior.  A user would have to have
a max attempts of at least 6 before this started to affect backoff
times.