Add additional waiter config validation for matching error codes
We should validate that if an error matcher references an error code,
then the operation actually has a modeled error with a matching code.

This won't pass as is, so I'm adding this as a reference for now.
We likely won't be able to get all the service models updated
immediately, so we may need to allow list existing cases that are valid
but not fully modeled.