Optimizes iter_lines handling long lines
In the previous implementation of iter_lines, lines that spanned chunks
could cause a quadratic iteration of the line data. This could result in
a much slower read time for files with long lines or lines significantly
longer than the chunk size used when reading them.

To avoid the issue, the StreamingBody now collects the line data and
only constructs the line (copying the bytes into a new buffer) when it
encounters a line ending.

Closes #2774